Convert navigation drawer to base activity
