+++ src/app/src/main/java/com/example/bookmark/BorrowedActivity.java
@@ -9,14 +9,9 @@ import android.view.View;
 import android.widget.AdapterView;
 import android.widget.ListView;
 
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.Toolbar;
-
 import com.example.bookmark.adapters.BookList;
 import com.example.bookmark.fragments.SearchDialogFragment;
 import com.example.bookmark.models.Book;
-import com.example.bookmark.util.DrawerProvider;
-import com.mikepenz.materialdrawer.Drawer;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -30,7 +25,8 @@ import java.util.List;
  *
  * @author Ryan Kortbeek.
  */
-public class BorrowedActivity extends AppCompatActivity implements SearchDialogFragment.OnFragmentInteractionListener {
+public class BorrowedActivity extends NavigationDrawerActivity
+    implements SearchDialogFragment.OnFragmentInteractionListener {
     public static final String EXTRA_BOOK = "com.example.bookmark.BOOK";
     public static final String SEARCHED_KEYWORDS = "com.example.bookmark" +
         ".SEARCH";
@@ -39,19 +35,12 @@ public class BorrowedActivity extends AppCompatActivity implements SearchDialogF
     private BookList borrowedBooksAdapter;
     private ListView borrowedBooksListView;
 
-    private Drawer navigationDrawer = null;
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setTheme(R.style.navigationActivity);
         setContentView(R.layout.activity_borrowed);
 
-        // setup toolbar
-        Toolbar toolbar = (Toolbar) findViewById(R.id.borrowed_toolbar);
-        setSupportActionBar(toolbar);
         getSupportActionBar().setTitle("Borrowed");
-        navigationDrawer = DrawerProvider.getDrawer(this, toolbar);
 
         borrowedBooksListView = findViewById(R.id.borrowed_books_listview);
 
@@ -86,16 +75,6 @@ public class BorrowedActivity extends AppCompatActivity implements SearchDialogF
         MenuInflater inflater = getMenuInflater();
         inflater.inflate(R.menu.menu_search, menu);
         return true;
-    }
-
-    @Override
-    public void onBackPressed() {
-        //handle the back press :D close the drawer first and if the drawer is closed close the activity
-        if (navigationDrawer != null && navigationDrawer.isDrawerOpen()) {
-            navigationDrawer.closeDrawer();
-        } else {
-            super.onBackPressed();
-        }
     }
 
     public boolean onOptionsItemSelected(MenuItem item) {
