+++ src/app/src/main/java/com/example/bookmark/ExploreActivity.java
@@ -7,16 +7,11 @@ import android.view.MenuInflater;
 import android.view.MenuItem;
 import android.widget.ListView;
 
-import androidx.appcompat.app.AppCompatActivity;
-import androidx.appcompat.widget.Toolbar;
-
 import com.example.bookmark.adapters.BookList;
 import com.example.bookmark.fragments.SearchDialogFragment;
 import com.example.bookmark.models.Book;
 import com.example.bookmark.server.FirebaseProvider;
 import com.example.bookmark.util.DialogUtil;
-import com.example.bookmark.util.DrawerProvider;
-import com.mikepenz.materialdrawer.Drawer;
 
 import java.util.ArrayList;
 import java.util.List;
@@ -29,7 +24,8 @@ import java.util.List;
  *
  * @author Ryan Kortbeek.
  */
-public class ExploreActivity extends AppCompatActivity implements SearchDialogFragment.OnFragmentInteractionListener {
+public class ExploreActivity extends NavigationDrawerActivity
+    implements SearchDialogFragment.OnFragmentInteractionListener {
     public static final String EXTRA_BOOK = "com.example.bookmark.BOOK";
     public static final String SEARCHED_KEYWORDS = "com.example.bookmark" +
         ".SEARCH";
@@ -38,19 +34,12 @@ public class ExploreActivity extends AppCompatActivity implements SearchDialogFr
     private BookList searchResultsAdapter;
     private ListView searchResultsListView;
 
-    private Drawer navigationDrawer = null;
-
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
-        setTheme(R.style.navigationActivity);
         setContentView(R.layout.activity_explore);
 
-        // setup toolbar
-        Toolbar toolbar = findViewById(R.id.explore_toolbar);
-        setSupportActionBar(toolbar);
         getSupportActionBar().setTitle("Explore");
-        navigationDrawer = DrawerProvider.getDrawer(this, toolbar);
 
         searchResultsListView = findViewById(R.id.search_results_listview);
 
@@ -77,16 +66,6 @@ public class ExploreActivity extends AppCompatActivity implements SearchDialogFr
         MenuInflater inflater = getMenuInflater();
         inflater.inflate(R.menu.menu_filter_search, menu);
         return true;
-    }
-
-    @Override
-    public void onBackPressed() {
-        //handle the back press :D close the drawer first and if the drawer is closed close the activity
-        if (navigationDrawer != null && navigationDrawer.isDrawerOpen()) {
-            navigationDrawer.closeDrawer();
-        } else {
-            super.onBackPressed();
-        }
     }
 
     public boolean onOptionsItemSelected(MenuItem item) {
