+++ src/app/src/main/res/layout/activity_my_books.xml
@@ -7,16 +7,6 @@
     android:focusable="true"
     tools:context=".MyBooksActivity">
 
-    <androidx.appcompat.widget.Toolbar
-        android:id="@+id/my_books_toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="?attr/actionBarSize"
-        android:background="?attr/colorPrimary"
-        android:elevation="4dp"
-        android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar"
-        app:layout_constraintTop_toTopOf="parent"
-        app:popupTheme="@style/ThemeOverlay.AppCompat.Light" />
-
     <ListView
         android:id="@+id/my_books_listview"
         android:layout_width="0dp"
@@ -24,11 +14,10 @@
         android:layout_marginStart="4dp"
         android:layout_marginTop="4dp"
         android:layout_marginEnd="4dp"
-        android:layout_marginBottom="4dp"
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/my_books_toolbar">
+        app:layout_constraintTop_toTopOf="parent">
 
     </ListView>
