+++ src/app/src/main/res/layout/activity_pending_requests.xml
@@ -7,16 +7,6 @@
     android:focusable="true"
     tools:context=".PendingRequestsActivity">
 
-    <androidx.appcompat.widget.Toolbar
-        android:id="@+id/requested_books_toolbar"
-        android:layout_width="match_parent"
-        android:layout_height="?attr/actionBarSize"
-        android:background="?attr/colorPrimary"
-        android:elevation="4dp"
-        android:theme="@style/ThemeOverlay.AppCompat.Dark.ActionBar"
-        app:layout_constraintTop_toTopOf="parent"
-        app:popupTheme="@style/ThemeOverlay.AppCompat.Light" />
-
     <ListView
         android:id="@+id/requested_books_listview"
         android:layout_width="0dp"
@@ -27,7 +17,7 @@
         app:layout_constraintBottom_toBottomOf="parent"
         app:layout_constraintEnd_toEndOf="parent"
         app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toBottomOf="@+id/requested_books_toolbar">
+        app:layout_constraintTop_toTopOf="parent">
 
     </ListView>
